wip: major refactoring
